Refactor: remove unused imports, expand * imports and rename classes
